# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface Testimonial {
  quote: string;
  attribution: string;
}

export interface ResolvedSettings {
  phone: string;
  email: string;
  phoneHref: string;
  mailtoHref: string;
  heroHeadline: string;
  heroSubhead: string;
  heroStyle: string;
  heroVideoUrl: string;
  testimonials: Testimonial[];
}

export interface RawSettings {
  phone: string;
  email: string;
  heroHeadline: string;
  heroSubhead: string;
  heroStyle: string;
  heroVideoUrl: string;
  testimonials: Testimonial[];
}

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  zip?: string;
  projectType?: string;
  message?: string;
}

export interface LeadUpdate {
  status?: string;
  notes?: string;
}

export interface LeadRecord {
  id: string;
  name: string;
  email: string;
  phone: string;
  zip: string;
  projectType: string;
  message: string;
  status: string;
  notes: string;
  createdAt: string;
  dateLabel: string;
}

export interface LeadStats {
  total: number;
  newCount: number;
  thisWeek: number;
  topProject: string;
}

export interface LeadsResponse {
  leads: LeadRecord[];
  stats: LeadStats;
}

export interface LeadUpdateResponse {
  ok: boolean;
  lead: LeadRecord;
}

export interface LoginInput {
  password: string;
}

export interface SettingsInput {
  phone: string;
  email: string;
  heroHeadline: string;
  heroSubhead: string;
  heroStyle: string;
  heroVideoUrl: string;
  testimonials: Testimonial[];
}