# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  zip?: string;
  projectType?: string;
  message?: string;
}