# lib/api-zod/src/generated/types/leadRecord.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadRecord {
  id: string;
  name: string;
  email: string;
  phone: string;
  zip: string;
  projectType: string;
  message: string;
  status: string;
  notes: string;
  createdAt: string;
  dateLabel: string;
}