# lib/api-zod/src/generated/types/leadStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadStats {
  total: number;
  newCount: number;
  thisWeek: number;
  topProject: string;
}