# lib/api-zod/src/generated/types/leadUpdateResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadRecord } from './leadRecord';

export interface LeadUpdateResponse {
  ok: boolean;
  lead: LeadRecord;
}