# lib/api-zod/src/generated/types/leadsResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadRecord } from './leadRecord';
import type { LeadStats } from './leadStats';

export interface LeadsResponse {
  leads: LeadRecord[];
  stats: LeadStats;
}