# lib/api-zod/src/generated/types/rawSettings.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * U Design For The Home — API
 * OpenAPI spec version: 0.1.0
 */
import type { Testimonial } from './testimonial';

export interface RawSettings {
  phone: string;
  email: string;
  heroHeadline: string;
  heroSubhead: string;
  heroStyle: string;
  heroVideoUrl: string;
  testimonials: Testimonial[];
}